Lungs-Pnemonia-Classification-project/streamlit_lungs/requirements.txt
streamlit
numpy
pillow
tensorflow